requirements.txt
python-telegram-bot==20.7
gigachat==0.1.12
python-dotenv==1.0.0